% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s3-package.R
\docType{package}
\name{s3-package}
\alias{s3}
\alias{s3-package}
\title{s3: Download Files from 'AWS S3'}
\description{
Download files hosted on 'AWS S3' (Amazon Web Services Simple Storage Service; \url{https://aws.amazon.com/s3/}) to a local directory based on their URI. Avoid downloading files that are already present locally. Allow for customization of where to store downloaded files.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/brokamp-group/s3}
  \item Report bugs at \url{https://github.com/brokamp-group/s3/issues}
}

}
\author{
\strong{Maintainer}: Cole Brokamp \email{cole.brokamp@gmail.com} (\href{https://orcid.org/0000-0002-0289-3151}{ORCID})

Authors:
\itemize{
  \item Erika Rasnick \email{erika.rasnick@cchmc.org}
  \item Robrecht Cannoodt
}

}
\keyword{internal}
