% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s3_operations.R
\name{s3_delete_bucket_lifecycle}
\alias{s3_delete_bucket_lifecycle}
\title{Deletes the lifecycle configuration from the bucket}
\usage{
s3_delete_bucket_lifecycle(Bucket)
}
\arguments{
\item{Bucket}{[required]}
}
\description{
Deletes the lifecycle configuration from the bucket.
}
\section{Request syntax}{
\preformatted{s3$delete_bucket_lifecycle(
  Bucket = "string"
)
}
}

\examples{
# The following example deletes lifecycle configuration on a bucket.
\donttest{s3$delete_bucket_lifecycle(
  Bucket = "examplebucket"
)}

}
\keyword{internal}
