% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s3_operations.R
\name{s3_delete_bucket_policy}
\alias{s3_delete_bucket_policy}
\title{Deletes the policy from the bucket}
\usage{
s3_delete_bucket_policy(Bucket)
}
\arguments{
\item{Bucket}{[required]}
}
\description{
Deletes the policy from the bucket.
}
\section{Request syntax}{
\preformatted{s3$delete_bucket_policy(
  Bucket = "string"
)
}
}

\examples{
# The following example deletes bucket policy on the specified bucket.
\donttest{s3$delete_bucket_policy(
  Bucket = "examplebucket"
)}

}
\keyword{internal}
