% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s3_operations.R
\name{s3_delete_bucket_website}
\alias{s3_delete_bucket_website}
\title{This operation removes the website configuration from the bucket}
\usage{
s3_delete_bucket_website(Bucket)
}
\arguments{
\item{Bucket}{[required]}
}
\description{
This operation removes the website configuration from the bucket.
}
\section{Request syntax}{
\preformatted{s3$delete_bucket_website(
  Bucket = "string"
)
}
}

\examples{
# The following example deletes bucket website configuration.
\donttest{s3$delete_bucket_website(
  Bucket = "examplebucket"
)}

}
\keyword{internal}
