% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s3_operations.R
\name{s3_get_bucket_lifecycle}
\alias{s3_get_bucket_lifecycle}
\title{No longer used, see the GetBucketLifecycleConfiguration operation}
\usage{
s3_get_bucket_lifecycle(Bucket)
}
\arguments{
\item{Bucket}{[required]}
}
\description{
No longer used, see the GetBucketLifecycleConfiguration operation.
}
\section{Request syntax}{
\preformatted{s3$get_bucket_lifecycle(
  Bucket = "string"
)
}
}

\examples{
# The following example gets ACL on the specified bucket.
\donttest{s3$get_bucket_lifecycle(
  Bucket = "acl1"
)}

}
\keyword{internal}
