% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s3_operations.R
\name{s3_get_bucket_lifecycle_configuration}
\alias{s3_get_bucket_lifecycle_configuration}
\title{Returns the lifecycle configuration information set on the bucket}
\usage{
s3_get_bucket_lifecycle_configuration(Bucket)
}
\arguments{
\item{Bucket}{[required]}
}
\description{
Returns the lifecycle configuration information set on the bucket.
}
\section{Request syntax}{
\preformatted{s3$get_bucket_lifecycle_configuration(
  Bucket = "string"
)
}
}

\examples{
# The following example retrieves lifecycle configuration on set on a
# bucket.
\donttest{s3$get_bucket_lifecycle_configuration(
  Bucket = "examplebucket"
)}

}
\keyword{internal}
