% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s3_operations.R
\name{s3_get_bucket_logging}
\alias{s3_get_bucket_logging}
\title{Returns the logging status of a bucket and the permissions users have to
view and modify that status}
\usage{
s3_get_bucket_logging(Bucket)
}
\arguments{
\item{Bucket}{[required]}
}
\description{
Returns the logging status of a bucket and the permissions users have to
view and modify that status. To use GET, you must be the bucket owner.
}
\section{Request syntax}{
\preformatted{s3$get_bucket_logging(
  Bucket = "string"
)
}
}

\keyword{internal}
