% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s3_operations.R
\name{s3_get_bucket_policy}
\alias{s3_get_bucket_policy}
\title{Returns the policy of a specified bucket}
\usage{
s3_get_bucket_policy(Bucket)
}
\arguments{
\item{Bucket}{[required]}
}
\description{
Returns the policy of a specified bucket.
}
\section{Request syntax}{
\preformatted{s3$get_bucket_policy(
  Bucket = "string"
)
}
}

\examples{
# The following example returns bucket policy associated with a bucket.
\donttest{s3$get_bucket_policy(
  Bucket = "examplebucket"
)}

}
\keyword{internal}
