% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s3_get_files.R
\name{s3_get_files}
\alias{s3_get_files}
\title{download several s3 files}
\usage{
s3_get_files(
  s3_uri,
  region = "us-east-2",
  download_folder = getOption("s3.download_folder", fs::path_wd("s3_downloads")),
  progress = FALSE,
  force = FALSE,
  confirm = TRUE,
  public = FALSE
)
}
\arguments{
\item{s3_uri}{vector of S3 object URIs}

\item{region}{AWS region for bucket containing the file (defaults to "us-east-2", but only required for private files)}

\item{download_folder}{location to download S3 objects}

\item{progress}{show download progress for each individual file? (currently only for public objects)}

\item{force}{force download to overwrite existing S3 objects}

\item{confirm}{ask user to interactively confirm downloads? (only possible when session is interactive)}

\item{public}{defaults to FALSE; if TRUE, ignore any environment
variables specifying AWS credentials and
attempt to download the file as publicly available}
}
\value{
tibble with s3_uris and corresponding file paths to downloaded files (invisibly)
}
\description{
download several s3 files
}
\details{
Progress messages for downloading several S3 objects at once cannot be silenced.

Like s3_get, S3 objects that already exists within the download_folder will not be re downloaded

Invisibly returning the S3 object file paths allows for further usage of files without hard coding.
(See example)
}
\examples{
\donttest{
s3_get_files(c(
    "s3://geomarker/testing_downloads/mtcars.rds",
    "s3://geomarker/testing_downloads/mtcars.fst"
))

dl_results <- s3_get_files(c(
    "s3://geomarker/testing_downloads/mtcars.rds",
    "s3://geomarker/testing_downloads/mtcars_again.rds"
))
lapply(dl_results$file_path, readRDS)

# download some larger files
s3_get_files(s3_uri = c(
    "s3://geomarker/testing_downloads/zctas_2000_contig_us_5072.rds",
    "s3://geomarker/testing_downloads/county_fips_contig_us_5072.rds"
))
}
}
