% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s3_operations.R
\name{s3_get_object_retention}
\alias{s3_get_object_retention}
\title{Retrieves an object's retention settings}
\usage{
s3_get_object_retention(Bucket, Key, VersionId, RequestPayer)
}
\arguments{
\item{Bucket}{[required] The bucket containing the object whose retention settings you want to
retrieve.}

\item{Key}{[required] The key name for the object whose retention settings you want to
retrieve.}

\item{VersionId}{The version ID for the object whose retention settings you want to
retrieve.}

\item{RequestPayer}{}
}
\description{
Retrieves an object's retention settings.
}
\section{Request syntax}{
\preformatted{s3$get_object_retention(
  Bucket = "string",
  Key = "string",
  VersionId = "string",
  RequestPayer = "requester"
)
}
}

\keyword{internal}
