% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s3_operations.R
\name{s3_head_bucket}
\alias{s3_head_bucket}
\title{This operation is useful to determine if a bucket exists and you have
permission to access it}
\usage{
s3_head_bucket(Bucket)
}
\arguments{
\item{Bucket}{[required]}
}
\description{
This operation is useful to determine if a bucket exists and you have
permission to access it.
}
\section{Request syntax}{
\preformatted{s3$head_bucket(
  Bucket = "string"
)
}
}

\examples{
# This operation checks to see if a bucket exists.
\donttest{s3$head_bucket(
  Bucket = "acl1"
)}

}
\keyword{internal}
