% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s3_get.R
\name{s3_get}
\alias{s3_get}
\title{download s3 file}
\usage{
s3_get(
  s3_uri,
  region = "us-east-2",
  download_folder = getOption("s3.download_folder", fs::path_wd("s3_downloads")),
  quiet = FALSE,
  progress = FALSE,
  force = FALSE,
  public = FALSE
)
}
\arguments{
\item{s3_uri}{URI for an S3 object}

\item{region}{AWS region for bucket containing the file (defaults to "us-east-2", but only required for private files)}

\item{download_folder}{location to download S3 object}

\item{quiet}{suppress messages?}

\item{progress}{show download progress? (currently only for public objects)}

\item{force}{force download to overwrite existing S3 object}

\item{public}{defaults to FALSE; if TRUE, ignore any environment
variables specifying AWS credentials and
attempt to download the file as publicly available}
}
\value{
file path to downloaded file (invisibly)
}
\description{
download s3 file
}
\details{
s3_get will politely refuse to download an S3 object if it already exists within the download_folder.

Invisibly returning the S3 object file path allows for further usage of file without hard coding.
(See example)
}
\examples{
\donttest{
s3_get(s3_uri = "s3://geomarker/testing_downloads/mtcars.rds")
s3_get("s3://geomarker/testing_downloads/mtcars.rds") |>
    readRDS()
}
}
