% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s3_get_files.R
\name{s3_get_files}
\alias{s3_get_files}
\title{download several s3 files}
\usage{
s3_get_files(
  s3_uri,
  download_folder = getOption("s3.download_folder", fs::path_wd("s3_downloads")),
  quiet = TRUE,
  force = FALSE,
  confirm = TRUE
)
}
\arguments{
\item{s3_uri}{vector of S3 object URIs}

\item{download_folder}{location to download S3 objects}

\item{quiet}{suppress individual download messages from s3_get?}

\item{force}{force download to overwrite existing S3 objects}

\item{confirm}{ask user to interactively confirm downloads?}
}
\value{
file paths to downloaded files (invisibly)
}
\description{
download several s3 files
}
\details{
Progress messages for downloading several S3 objects cannot be silenced.

Like s3_get, S3 objects that already exists within the download_folder will not be re downloaded

Invisibly returning the S3 object file path allows for further usage of file without hard coding.
(See example)
}
\examples{
s3_get_files(c("s3://geomarker/testing_downloads/mtcars.rds",
               "s3://geomarker/testing_downloads/mtcars.fst"))
}
